backend/requirements.txt
# ── Core ──────────────────────────────────
Flask==3.0.3
python-dotenv==1.0.1

# ── Database ──────────────────────────────
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Flask-Migrate==4.0.7

# ── Auth ──────────────────────────────────
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0

# ── Validation ────────────────────────────
marshmallow==3.21.3
Flask-Marshmallow==1.2.1

# ── Email ─────────────────────────────────
Flask-Mail==0.10.0

# ── Payment ───────────────────────────────
razorpay==1.4.2

# ── CORS ──────────────────────────────────
Flask-Cors==4.0.1

# ── Utils ─────────────────────────────────
Pillow==10.4.0
python-slugify==8.0.4

anthropic>=0.20.0